# apps/api/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "hiphouse API",
    "description": "크롤러가 채운 Postgres(artists/albums) CRUD 서버.",
    "version": "0.1.0"
  },
  "servers": [{ "url": "/" }],
...
  "paths": {
    "/albums": {
      "get": {
        "tags": ["albums"],
        "summary": "앨범 목록 (연도/아티스트 필터)",
        "parameters": [
          { "name": "year", "in": "query", "schema": { "type": "integer" }, "description": "발매 연도로 필터" },
          { "name": "artist_id", "in": "query", "schema": { "type": "string" }, "description": "아티스트 ID로 필터" },
          { "name": "q", "in": "query", "schema": { "type": "string" }, "description": "검색어 — 앨범명(한글 표시 이름 포함), 크레딧 아티스트명(한글 표시 이름 포함), 아티스트 연관검색어(aliases), 트랙명(한글 표시 이름 포함)에 매칭" },
          { "name": "deleted", "in": "query", "schema": { "type": "string", "enum": ["only"] }, "description": "only = 소프트삭제된 앨범만 (관리자 전용; 비관리자에겐 무시됨)" },
          { "name": "limit", "in": "query", "schema": { "type": "integer", "default": 50, "minimum": 1, "maximum": 200 } },
          { "name": "offset", "in": "query", "schema": { "type": "integer", "default": 0, "minimum": 0 } }
        ],
        "responses": {
          "200": {
            "description": "앨범 배열",
            "content": { "application/json": { "schema": { "type": "array", "items": { "$ref": "#/components/schemas/Album" } } } }
          }
        }
      },
      "post": {
        "tags": ["albums"],
        "summary": "앨범 생성",
        "requestBody": {
          "required": true,
          "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Album" } } }
        },
        "responses": {
          "201": { "description": "생성됨", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Album" } } } },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "409": { "description": "id 중복", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } }
        }
      }
    },
    "/albums/{id}": {
      "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }],
      "get": {
        "tags": ["albums"],
        "summary": "앨범 단건 조회",
        "responses": {
          "200": { "description": "앨범", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Album" } } } },
          "404": { "$ref": "#/components/responses/NotFound" }
        }
      },
      "put": {
        "tags": ["albums"],
        "summary": "앨범 수정 (전체 교체)",
        "requestBody": {
          "required": true,
          "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Album" } } }
        },
        "responses": {
          "200": { "description": "수정됨", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Album" } } } },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "404": { "$ref": "#/components/responses/NotFound" }
        }
      },
      "delete": {
        "tags": ["albums"],
        "summary": "앨범 삭제",
        "responses": {
          "204": { "description": "삭제됨" },
          "404": { "$ref": "#/components/responses/NotFound" }
        }
      }
    },
...
  },
  "components": {
    "responses": {
      "BadRequest": { "description": "잘못된 요청", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } },
      "NotFound": { "description": "없음", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } }
    },
    "schemas": {
      "Album": {
        "type": "object",
        "required": ["id", "name", "artists"],
        "properties": {
          "id": { "type": "string", "example": "4aawyAB9vmqN3uQ7FjRGTy" },
          "name": { "type": "string", "example": "Album Name" },
          "display_name": { "type": "string", "nullable": true, "example": "새삥", "description": "관리자가 등록한 한글 표시 이름. 없으면 name으로 표시하고, 검색 시 name과 함께 매칭됨" },
          "artists": {
            "type": "array",
            "description": "크레딧된 모든 아티스트 (Spotify 순서). album_artists 조인에서 집계.",
            "items": { "$ref": "#/components/schemas/Artist" }
          },
          "release_date": { "type": "string", "nullable": true, "example": "2024-05-01" },
          "year": { "type": "integer", "nullable": true, "example": 2024 },
          "album_type": { "type": "string", "nullable": true, "example": "album" },
          "total_tracks": { "type": "integer", "nullable": true, "example": 12 },
          "image_url": { "type": "string", "nullable": true },
          "spotify_url": { "type": "string", "nullable": true },
          "apple_id": { "type": "string", "nullable": true, "description": "Apple Music 카탈로그 앨범 id — UPC 매칭으로 연결 (apple.go). 미연결이면 null" },
          "label": { "type": "string", "nullable": true, "description": "Apple recordLabel. Spotify label 필드는 2026-02 개편으로 죽어 Apple이 유일한 소스" },
          "genres": { "type": "array", "nullable": true, "items": { "type": "string" }, "description": "Apple 앨범 장르 (루트 \"음악\" 제외)" },
          "copyright": { "type": "string", "nullable": true, "description": "℗ 한 줄 (Apple copyright)" },
          "content_rating": { "type": "string", "nullable": true, "description": "Apple contentRating: explicit | clean | null" },
          "editorial_notes": { "type": "string", "nullable": true, "description": "Apple 에디토리얼 노트 평문 (태그 제거, 줄바꿈 포함)" }
        }
      },
      "Artist": {
        "type": "object",
        "required": ["id"],
        "properties": {
          "id": { "type": "string", "example": "1uNFoZAHBGtllmzznpCI3s" },
          "name": { "type": "string", "nullable": true, "example": "Artist Name" },
          "display_name": { "type": "string", "nullable": true, "example": "지코", "description": "관리자가 등록한 한글 표시 이름. 없으면 name으로 표시하고, 검색 시 name과 함께 매칭됨" },
          "image_url": { "type": "string", "nullable": true, "description": "Spotify 아티스트 사진 (없을 수 있음)" },
          "genres": { "type": "array", "items": { "type": "string" }, "nullable": true, "description": "Spotify 아티스트 장르 (비어있을 수 있음)" },
          "spotify_url": { "type": "string", "nullable": true, "description": "Spotify 아티스트 프로필 URL" },
          "aliases": { "type": "array", "items": { "type": "string" }, "nullable": true, "description": "연관검색어 — 관리자가 등록한 검색 키워드 (예: 영문명의 한글 표기). 검색 시 name과 함께 매칭됨" },
          "followers": { "type": "integer", "nullable": true, "description": "Spotify 팔로워 수 — 크롤러/enrich가 채움 (아직 수집 전이면 null). 화면 표시용이 아니라 관리자 enrich 판단용" },
          "releases_watch": { "type": "boolean", "description": "신보 감시 대상 여부 — 신보 체크가 이 플래그 켜진 아티스트만 확인. 관리자 토글 소유" },
          "follower_count": { "type": "integer", "description": "서비스 내부 팔로워 수 — follows 테이블 집계" }
        }
      },
...
      "Error": {
        "type": "object",
        "properties": { "error": { "type": "string" } }
      }
    }
  }
}